+++ packages/wrangler/src/__tests__/helpers/mock-legacy-script.ts
@@ -1,0 +1,23 @@
+import { http, HttpResponse } from "msw";
+import { msw } from "./msw";
+
+type LegacyScriptInfo = { id: string; migration_tag?: string };
+
+export function mockLegacyScriptData(options: { scripts: LegacyScriptInfo[] }) {
+	const { scripts } = options;
+	msw.use(
+		http.get(
+			"*/accounts/:accountId/workers/scripts",
+			({ params }) => {
+				expect(params.accountId).toEqual("some-account-id");
+				return HttpResponse.json({
+					success: true,
+					errors: [],
+					messages: [],
+					result: scripts,
+				});
+			},
+			{ once: true }
+		)
+	);
+}
